Add files via upload
 Console-Based-Calculator-Application-Using-C-Programming -
The proposed project is a simple consoled-based computer application in C programing language that performs the basic operations of addition, subtraction, multiplication, and division. 
***Functionality***
calculator application that accepts two input numbers and performs an arithmetic operation such as addition, subtraction, division, and multiplication, then prints the output on the console, so that it can assist me in do mathematics faster. The simple calculator application will have the following operands +, -, *, /, and =. In any situation, the calculator app has to produce a correct result defined by the well-known arithmetic rules. 
After successfully performing the basic operation, the functionality of the calculator need to expand to perform higher functions and warn the user of various errors. If the calculation is impossible the calculator has to display information helping the user resolve the erroneous situation, for example, on encountering division by 0 the output should display “Cannot divide by zero”.